Move ColliderStore to a separate module
