+++ src/collisions/colliders.rs
@@ -6,13 +6,10 @@ use bevy::{
 };
 
 use super::{
-    circle_to_circle, rect_to_circle, rect_to_rect,
-    shapes::{ColliderShape, ColliderShapeType},
-    spatial_hash::SpatialHash,
-    ColliderId, CollisionResultRef, RaycastHit,
+    circle_to_circle, rect_to_circle, rect_to_rect, shapes::{ColliderShape, ColliderShapeType}, spatial_hash::SpatialHash, store::{ColliderStore, ALL_LAYERS}, ColliderId, CollisionResultRef, RaycastHit
 };
 
-pub const ALL_LAYERS: i32 = -1;
+
 
 #[derive(Debug, Clone, PartialEq, Reflect)]
 pub struct Collider {
@@ -257,202 +254,6 @@ pub struct ColliderBundle {
     pub collider: ColliderComponent,
 }
 
-pub trait ColliderIdResolver {
-    fn get(&self, id: ColliderId) -> Option<&Collider>;
-    fn get_mut(&mut self, id: ColliderId) -> Option<&mut Collider>;
-}
-
-static COLLIDER_ID_GEN: AtomicU32 = AtomicU32::new(0);
-
-#[derive(Resource)]
-pub struct ColliderStore {
-    pub colliders: HashMap<ColliderId, Collider>,
-    spatial_hash: SpatialHash,
-}
-
-impl Default for ColliderStore {
-    fn default() -> Self {
-        Self {
-            colliders: HashMap::new(),
-            spatial_hash: SpatialHash::new(100),
-        }
-    }
-}
-
-impl ColliderStore {
-    pub fn new(cell_size: i32) -> Self {
-        Self {
-            spatial_hash: SpatialHash::new(cell_size),
-            ..default()
-        }
-    }
-
-    pub fn create_and_register(&mut self, shape_type: ColliderShapeType) -> ColliderComponent {
-        let collider = Collider::new(shape_type);
-        let id = self.register(collider);
-
-        ColliderComponent { id }
-    }
-
-    pub fn register(&mut self, mut collider: Collider) -> ColliderId {
-        let id = COLLIDER_ID_GEN.fetch_add(1, Ordering::SeqCst);
-        let id = ColliderId(id);
-        collider.id = id;
-
-        self.colliders.insert(id, collider);
-
-        id
-    }
-
-    pub fn from_component(&self, component: &ColliderComponent) -> Option<&Collider> {
-        self.colliders.get(&component.id)
-    }
-
-    pub fn from_component_mut(&mut self, component: &ColliderComponent) -> Option<&mut Collider> {
-        self.colliders.get_mut(&component.id)
-    }
-
-    pub fn remove(&mut self, component: ColliderComponent) -> Option<Collider> {
-        let col = self.colliders.get_mut(&component.id);
-        if col.is_none() {
-            return None;
-        }
-
-        let col: &mut Collider = col.unwrap();
-        col.is_registered = false;
-
-        self.spatial_hash.remove(&col);
-        self.colliders.remove(&component.id)
-    }
-
-    pub fn get_as_components(&self) -> Vec<ColliderComponent> {
-        let mut res = vec![];
-
-        for (id, _v) in &self.colliders {
-            res.push(ColliderComponent { id: *id });
-        }
-
-        res
-    }
-
-    pub fn aabb_broadphase(
-        &self,
-        rect: super::Rect,
-        layer_mask: Option<i32>,
-    ) -> HashSet<ColliderId> {
-        let layer_mask = match layer_mask {
-            Some(val) => val,
-            None => ALL_LAYERS,
-        };
-
-        self.spatial_hash
-            .aabb_broadphase(&rect, None, layer_mask, |id| self.colliders.get(id))
-    }
-
-    pub fn aabb_broadphase_excluding_self(
-        &self,
-        self_collider: ColliderId,
-        rect: super::Rect,
-        layer_mask: Option<i32>,
-    ) -> HashSet<ColliderId> {
-        let layer_mask = match layer_mask {
-            Some(val) => val,
-            None => ALL_LAYERS,
-        };
-
-        self.spatial_hash
-            .aabb_broadphase(&rect, Some(self_collider), layer_mask, |id| {
-                self.colliders.get(id)
-            })
-    }
-
-    pub fn linecast(
-        &self,
-        start: Vec2,
-        end: Vec2,
-        layer_mask: Option<i32>,
-    ) -> (i32, Vec<RaycastHit>) {
-        let layer_mask = match layer_mask {
-            Some(val) => val,
-            None => ALL_LAYERS,
-        };
-
-        self.spatial_hash
-            .linecast(start, end, layer_mask, |id| self.colliders.get(id))
-    }
-
-    pub fn overlap_circle(&self, circle_center: Vec2, radius: f32, layer_mask: Option<i32>) -> i32 {
-        let layer_mask = match layer_mask {
-            Some(val) => val,
-            None => ALL_LAYERS,
-        };
-
-        let mut results = vec![ColliderId(0); 1];
-
-        self.spatial_hash
-            .overlap_circle(circle_center, radius, &mut results, layer_mask, |id| {
-                self.colliders.get(id)
-            })
-    }
-
-    pub fn overlap_rectangle(&self, rect: super::Rect, layer_mask: Option<i32>) -> i32 {
-        let layer_mask = match layer_mask {
-            Some(val) => val,
-            None => ALL_LAYERS,
-        };
-
-        let mut results = vec![ColliderId(0); 1];
-
-        self.spatial_hash
-            .overlap_rectangle(&rect, &mut results, layer_mask, |id| self.colliders.get(id))
-    }
-
-    pub(crate) fn clear_hash(&mut self) {
-        self.spatial_hash.clear();
-    }
-
-    pub(crate) fn update_single(&mut self, component: &ColliderComponent, transform: &Transform) {
-        if let Some(col) = self.colliders.get(&component.id) {
-            if col.is_registered {
-                self.spatial_hash.remove(col);
-            }
-        }
-
-        if let Some(col) = self.get_mut(component.id) {
-            col.is_registered = true;
-            col.update_from_transform(transform);
-        }
-
-        if let Some(col) = self.colliders.get(&component.id) {
-            self.spatial_hash.register(col);
-        }
-    }
-
-    pub(crate) fn added_with_transform(
-        &mut self,
-        component: &ColliderComponent,
-        transform: &Transform,
-    ) {
-        if let Some(col) = self.get_mut(component.id) {
-            col.is_registered = true;
-            col.update_from_transform(transform);
-        }
-
-        if let Some(col) = self.colliders.get(&component.id) {
-            self.spatial_hash.register(col);
-        }
-    }
-}
-
-impl ColliderIdResolver for ColliderStore {
-    fn get(&self, id: ColliderId) -> Option<&Collider> {
-        self.colliders.get(&id)
-    }
-
-    fn get_mut(&mut self, id: ColliderId) -> Option<&mut Collider> {
-        self.colliders.get_mut(&id)
-    }
-}
 
 pub struct CollisionPlugin;
